requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.23.1
azure-storage-blob==12.26.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.1
chromadb==0.4.0
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.5
dataclasses-json==0.5.9
distro==1.9.0
emoji==2.14.1
fastapi==0.99.1
filelock==3.18.0
filetype==1.2.0
Flask==2.3.3
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
langchain==0.0.267
langdetect==1.0.9
langsmith==0.0.92
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-enum==1.5.1
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.3
mypy_extensions==1.1.0
nltk==3.9.1
numexpr==2.11.0
numpy==1.26.4
onnxruntime==1.22.1
openai==0.28.0
openapi-schema-pydantic==1.2.4
overrides==7.7.0
packaging==25.0
pandas==2.3.1
posthog==6.3.1
propcache==0.3.2
protobuf==6.31.1
pulsar-client==3.8.0
pycparser==2.22
pydantic==1.10.22
PyJWT==2.10.1
pypdf==3.9.0
PyPika==0.48.9
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-iso639==2025.2.18
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.27.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.10.30
urllib3==2.0.7
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==2.3.7
yarl==1.20.1
